chore(amazonq): refactor inline completion codes (#2367)
