+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/contants/constants.ts
