+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/tracker/codeDiffTracker.test.ts
